Split main.go up into several files
