+++ main.go
@@ -1,20 +1,12 @@
 package main
 
 import (
-	"archive/tar"
-	"bytes"
-	"compress/gzip"
-	"crypto/md5"
 	"crypto/rand"
-	"crypto/sha1"
-	"crypto/sha256"
 	"encoding/base64"
-	"encoding/hex"
 	"encoding/json"
 	"errors"
 	"flag"
 	"fmt"
-	"io"
 	"io/ioutil"
 	"log"
 	"net/http"
@@ -24,13 +16,8 @@ import (
 	"sync"
 	"time"
 
-	"github.com/blakesmith/ar"
 	"github.com/boltdb/bolt"
 	"github.com/fsnotify/fsnotify"
-	"golang.org/x/crypto/openpgp"
-	"golang.org/x/crypto/openpgp/armor"
-	"golang.org/x/crypto/openpgp/clearsign"
-	"golang.org/x/crypto/openpgp/packet"
 )
 
 type conf struct {
@@ -49,13 +36,6 @@ type conf struct {
 
 func (c conf) ArchPath(distro, section, arch string) string {
 	return filepath.Join(c.RootRepoPath, "dists", distro, section, "binary-"+arch)
-}
-
-type deleteObj struct {
-	Filename   string
-	DistroName string
-	Arch       string
-	Section    string
 }
 
 var (
@@ -218,478 +198,6 @@ func createDirs(config conf) error {
 	return nil
 }
 
-func inspectPackage(filename string) (string, error) {
-	f, err := os.Open(filename)
-	if err != nil {
-		return "", fmt.Errorf("error opening package file %s: %s", filename, err)
-	}
-
-	if *verbose {
-		log.Printf("Inpecting package file \"%s\"", filename)
-	}
-
-	arReader := ar.NewReader(f)
-	defer f.Close()
-	var controlBuf bytes.Buffer
-
-	for {
-		header, err := arReader.Next()
-
-		if err == io.EOF {
-			break
-		}
-
-		if err != nil {
-			return "", fmt.Errorf("error in inspectPackage loop: %s", err)
-		}
-
-		if strings.TrimRight(header.Name, "/") == "control.tar.gz" {
-			io.Copy(&controlBuf, arReader)
-			if *verbose {
-				log.Println("\t Found a package control file")
-			}
-			return inspectPackageControl(controlBuf)
-		}
-
-	}
-	return "", nil
-}
-
-func inspectPackageControl(filename bytes.Buffer) (string, error) {
-	gzf, err := gzip.NewReader(bytes.NewReader(filename.Bytes()))
-	if err != nil {
-		return "", fmt.Errorf("error creating gzip reader: %s", err)
-	}
-
-	tarReader := tar.NewReader(gzf)
-	var controlBuf bytes.Buffer
-	for {
-		header, err := tarReader.Next()
-
-		if err == io.EOF {
-			break
-		}
-
-		if err != nil {
-			return "", fmt.Errorf("failed to inspect package: %s", err)
-		}
-
-		name := header.Name
-
-		switch header.Typeflag {
-		case tar.TypeDir:
-			continue
-		case tar.TypeReg:
-			if name == "./control" {
-				io.Copy(&controlBuf, tarReader)
-				return strings.TrimRight(controlBuf.String(), "\n") + "\n", nil
-			}
-		default:
-			log.Printf(
-				"Unable to figure out type : %c in file %s\n",
-				header.Typeflag, name,
-			)
-		}
-	}
-	return "", nil
-}
-
-func createPackagesGz(config conf, distro, section, arch string) error {
-
-	if *verbose {
-		log.Printf("Rebuilding Packages.gz file for %s %s %s", distro, section, arch)
-	}
-
-	packageFile, err := os.Create(filepath.Join(config.ArchPath(distro, section, arch), "Packages"))
-	if err != nil {
-		return fmt.Errorf("failed to create Packages: %s", err)
-	}
-	packageGzFile, err := os.Create(filepath.Join(config.ArchPath(distro, section, arch), "Packages.gz"))
-	if err != nil {
-		return fmt.Errorf("failed to create packages.gz: %s", err)
-	}
-	defer packageFile.Close()
-	defer packageGzFile.Close()
-	gzOut := gzip.NewWriter(packageGzFile)
-	defer gzOut.Close()
-
-	writer := io.MultiWriter(packageFile, gzOut)
-
-	// loop through each directory
-	// run inspectPackage
-	dirList, err := ioutil.ReadDir(config.ArchPath(distro, section, arch))
-	if err != nil {
-		return fmt.Errorf("scanning: %s: %s", config.ArchPath(distro, section, arch), err)
-	}
-	for i, debFile := range dirList {
-		if strings.HasSuffix(debFile.Name(), "deb") {
-			var packBuf bytes.Buffer
-			debPath := filepath.Join(config.ArchPath(distro, section, arch), debFile.Name())
-			tempCtlData, err := inspectPackage(debPath)
-			if err != nil {
-				return err
-			}
-			packBuf.WriteString(tempCtlData)
-			dir := filepath.ToSlash(filepath.Join("dists", distro, section, "binary-"+arch, debFile.Name()))
-			fmt.Fprintf(&packBuf, "Filename: %s\n", dir)
-			fmt.Fprintf(&packBuf, "Size: %d\n", debFile.Size())
-			f, err := os.Open(debPath)
-			if err != nil {
-				log.Println("error opening deb file: ", err)
-			}
-			defer f.Close()
-
-			var (
-				md5hash    = md5.New()
-				sha1hash   = sha1.New()
-				sha256hash = sha256.New()
-			)
-			_, err = io.Copy(io.MultiWriter(md5hash, sha1hash, sha256hash), f)
-			if err != nil {
-				return fmt.Errorf("Error hashing file for Packages file: %s", err)
-			}
-			fmt.Fprintf(&packBuf, "MD5sum: %s\n",
-				hex.EncodeToString(md5hash.Sum(nil)))
-			fmt.Fprintf(&packBuf, "SHA1: %s\n",
-				hex.EncodeToString(sha1hash.Sum(nil)))
-			fmt.Fprintf(&packBuf, "SHA256: %s\n",
-				hex.EncodeToString(sha256hash.Sum(nil)))
-			if i != (len(dirList) - 1) {
-				packBuf.WriteString("\n\n")
-			}
-			writer.Write(packBuf.Bytes())
-			f = nil
-		}
-	}
-
-	return nil
-}
-
-func createRelease(config conf, distro string) error {
-
-	if *verbose {
-		log.Printf("Creating release file for \"%s\"", distro)
-	}
-
-	workingDirectory := filepath.Join(config.RootRepoPath, "dists", distro)
-
-	outfile, err := os.Create(filepath.Join(workingDirectory, "Release"))
-	if err != nil {
-		return fmt.Errorf("failed to create Release: %s", err)
-	}
-	defer outfile.Close()
-
-	currentTime := Now().UTC()
-	fmt.Fprintf(outfile, "Suite: %s\n", distro)
-	fmt.Fprintf(outfile, "Codename: %s\n", distro)
-	fmt.Fprintf(outfile, "Components: %s\n", strings.Join(config.Sections, " "))
-	fmt.Fprintf(outfile, "Architectures: %s\n", strings.Join(config.SupportArch, " "))
-	fmt.Fprintf(outfile, "Date: %s\n", currentTime.Format("Mon, 02 Jan 2006 15:04:05 UTC"))
-
-	var md5Sums strings.Builder
-	var sha1Sums strings.Builder
-	var sha256Sums strings.Builder
-
-	err = filepath.Walk(workingDirectory, func(path string, file os.FileInfo, err error) error {
-		if err != nil {
-			return err
-		}
-
-		if strings.HasSuffix(path, "Packages.gz") || strings.HasSuffix(path, "Packages") {
-
-			var (
-				md5hash    = md5.New()
-				sha1hash   = sha1.New()
-				sha256hash = sha256.New()
-			)
-
-			relPath, _ := filepath.Rel(workingDirectory, path)
-			spath := filepath.ToSlash(relPath)
-			f, err := os.Open(path)
-
-			if err != nil {
-				log.Println("Error opening the Packages file for reading", err)
-			}
-
-			if _, err = io.Copy(io.MultiWriter(md5hash, sha1hash, sha256hash), f); err != nil {
-				return fmt.Errorf("Error hashing file for Release list: %s", err)
-			}
-			fmt.Fprintf(&md5Sums, " %s %d %s\n",
-				hex.EncodeToString(md5hash.Sum(nil)),
-				file.Size(), spath)
-			fmt.Fprintf(&sha1Sums, " %s %d %s\n",
-				hex.EncodeToString(sha1hash.Sum(nil)),
-				file.Size(), spath)
-			fmt.Fprintf(&sha256Sums, " %s %d %s\n",
-				hex.EncodeToString(sha256hash.Sum(nil)),
-				file.Size(), spath)
-
-			f = nil
-		}
-		return nil
-	})
-
-	if err != nil {
-		return fmt.Errorf("Error scanning for Packages files: %s", err)
-	}
-
-	outfile.WriteString("MD5Sum:\n")
-	outfile.WriteString(md5Sums.String())
-	outfile.WriteString("SHA1:\n")
-	outfile.WriteString(sha1Sums.String())
-	outfile.WriteString("SHA256:\n")
-	outfile.WriteString(sha256Sums.String())
-
-	if err = signRelease(config, outfile.Name()); err != nil {
-		return fmt.Errorf("Error signing Release file: %s", err)
-	}
-
-	return nil
-}
-
-func signRelease(config conf, filename string) error {
-
-	if *verbose {
-		log.Printf("Signing release file \"%s\"", filename)
-	}
-
-	entity := createEntityFromPrivateKey(config.PrivateKey)
-
-	workingDirectory := filepath.Dir(filename)
-
-	releaseFile, err := os.Open(filename)
-	if err != nil {
-		return fmt.Errorf("Error opening Release file (%s) for writing: %s", filename, err)
-	}
-
-	releaseGpg, err := os.Create(filepath.Join(workingDirectory, "Release.gpg"))
-	if err != nil {
-		return fmt.Errorf("Error creating Release.gpg file for writing: %s", err)
-	}
-	defer releaseGpg.Close()
-
-	err = openpgp.ArmoredDetachSign(releaseGpg, entity, releaseFile, nil)
-	if err != nil {
-		return fmt.Errorf("Error writing signature to Release.gpg file: %s", err)
-	}
-
-	releaseFile.Seek(0, 0)
-
-	inlineRelease, err := os.Create(filepath.Join(workingDirectory, "InRelease"))
-	if err != nil {
-		return fmt.Errorf("Error creating InRelease file for writing: %s", err)
-	}
-	defer inlineRelease.Close()
-
-	writer, err := clearsign.Encode(inlineRelease, entity.PrivateKey, nil)
-	if err != nil {
-		return fmt.Errorf("Error signing InRelease file : %s", err)
-	}
-
-	io.Copy(writer, releaseFile)
-	writer.Close()
-
-	return nil
-}
-
-func createKeyPair(name, comment, email string) (*openpgp.Entity, string, string) {
-
-	entity, err := openpgp.NewEntity(name, comment, email, nil)
-	if err != nil {
-		log.Fatalf("Error creating openpgp entity: %s", err)
-	}
-
-	serializedPrivateEntity := bytes.NewBuffer(nil)
-	entity.SerializePrivate(serializedPrivateEntity, nil)
-	serializedEntity := bytes.NewBuffer(nil)
-	entity.Serialize(serializedEntity)
-
-	buf := bytes.NewBuffer(nil)
-	headers := map[string]string{"Version": "GnuPG v1"}
-	w, err := armor.Encode(buf, openpgp.PublicKeyType, headers)
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	_, err = w.Write(serializedEntity.Bytes())
-	if err != nil {
-		log.Fatalf("Error encoding public key: %s", err)
-	}
-	w.Close()
-	publicKey := buf.String()
-
-	buf = bytes.NewBuffer(nil)
-	w, err = armor.Encode(buf, openpgp.PrivateKeyType, headers)
-	if err != nil {
-		log.Fatal(err)
-	}
-	_, err = w.Write(serializedPrivateEntity.Bytes())
-	if err != nil {
-		log.Fatalf("Error encoding private key: %s", err)
-	}
-	w.Close()
-	privateKey := buf.String()
-
-	return entity, publicKey, privateKey
-}
-
-func createEntityFromPrivateKey(privateKeyPath string) *openpgp.Entity {
-
-	privateKeyData, err := os.Open(privateKeyPath)
-
-	if err != nil {
-		log.Fatalf("Error opening private key file: %s", err)
-	}
-	defer privateKeyData.Close()
-
-	block, err := armor.Decode(privateKeyData)
-
-	if err != nil {
-		log.Fatalf("Error decoding private key data: %s", err)
-	}
-
-	if block.Type != openpgp.PrivateKeyType {
-		log.Fatalf("Invalid private key type %s", block.Type)
-	}
-
-	reader := packet.NewReader(block.Body)
-	pkt, err := reader.Next()
-
-	if err != nil {
-		log.Fatalf("Error reading private key data: %s", err)
-	}
-
-	privateKey, ok := pkt.(*packet.PrivateKey)
-	if !ok {
-		log.Fatalf("Error parsing private key")
-	}
-
-	e := openpgp.Entity{
-		PrivateKey: privateKey,
-	}
-
-	return &e
-}
-
-func createKeyHandler(workingDirectory, name, email string) {
-
-	_, publicKey, privateKey := createKeyPair(name, "Generated by deb-simple", email)
-
-	pubFile, err := os.Create(filepath.Join(workingDirectory, "public.key"))
-	if err != nil {
-		log.Fatalf("Could not open public key file for writing: %s", err)
-	}
-	defer pubFile.Close()
-
-	pubFile.WriteString(publicKey)
-	
-	priFile, err := os.Create(filepath.Join(workingDirectory, "private.key"))
-	if err != nil {
-		log.Fatalf("Could not open private key file for writing: %s", err)
-	}
-	defer priFile.Close()
-
-	priFile.WriteString(privateKey)
-}
-
-func uploadHandler(config conf, db *bolt.DB) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		log.Printf("%s %s %s\n", r.RemoteAddr, r.Method, r.URL)
-		if r.Method != "POST" {
-			http.Error(w, "method not supported", http.StatusMethodNotAllowed)
-			return
-		}
-		if config.EnableAPIKeys {
-			apiKey := r.URL.Query().Get("key")
-			if apiKey == "" {
-				http.Error(w, "api key not present", http.StatusUnauthorized)
-				return
-			}
-			if !validateAPIkey(db, apiKey) {
-				http.Error(w, "api key not valid", http.StatusUnauthorized)
-				return
-			}
-		}
-		archType := r.URL.Query().Get("arch")
-		if archType == "" {
-			archType = "all"
-		}
-		distroName := r.URL.Query().Get("distro")
-		if distroName == "" {
-			distroName = "stable"
-		}
-		section := r.URL.Query().Get("section")
-		if section == "" {
-			section = "main"
-		}
-		reader, err := r.MultipartReader()
-		if err != nil {
-			httpErrorf(w, "error creating multipart reader: %s", err)
-			return
-		}
-		for {
-			part, err := reader.NextPart()
-			if err == io.EOF {
-				break
-			}
-			if part.FileName() == "" {
-				continue
-			}
-			dst, err := os.Create(filepath.Join(config.ArchPath(distroName, section, archType), part.FileName()))
-
-			if *verbose {
-				log.Printf("Deb package %s has been uploaded to %s %s %s", part.FileName(), distroName, section, archType)
-			}
-
-			if err != nil {
-				httpErrorf(w, "error creating deb file: %s", err)
-				return
-			}
-			defer dst.Close()
-			if _, err := io.Copy(dst, part); err != nil {
-				httpErrorf(w, "error writing deb file: %s", err)
-				return
-			}
-		}
-		w.WriteHeader(http.StatusOK)
-	})
-}
-
-func deleteHandler(config conf, db *bolt.DB) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		log.Printf("%s %s %s\n", r.RemoteAddr, r.Method, r.URL)
-		if r.Method != "DELETE" {
-			http.Error(w, "method not supported", http.StatusMethodNotAllowed)
-			return
-		}
-		if config.EnableAPIKeys {
-			apiKey := r.URL.Query().Get("key")
-			if apiKey == "" {
-				http.Error(w, "api key not present", http.StatusUnauthorized)
-				return
-			}
-			if !validateAPIkey(db, apiKey) {
-				http.Error(w, "api key not valid", http.StatusUnauthorized)
-				return
-			}
-		}
-		var toDelete deleteObj
-		if err := json.NewDecoder(r.Body).Decode(&toDelete); err != nil {
-			httpErrorf(w, "failed to decode json: %s", err)
-			return
-		}
-		debPath := filepath.Join(config.ArchPath(toDelete.DistroName, toDelete.Section, toDelete.Arch), toDelete.Filename)
-		if err := os.Remove(debPath); err != nil {
-			httpErrorf(w, "failed to delete: %s", err)
-			return
-		}
-
-		if *verbose {
-			log.Printf("Deb package %s has been deleted", toDelete.Filename)
-		}
-	})
-}
-
 func openDB() *bolt.DB {
 	// open/create database for API keys
 	db, err := bolt.Open("debsimple.db", 0600, &bolt.Options{Timeout: 1 * time.Second})
@@ -721,25 +229,4 @@ func createAPIkey(db *bolt.DB) (string, error) {
 		return "", err
 	}
 	return apiKey, nil
-}
-
-func validateAPIkey(db *bolt.DB, key string) bool {
-	err := db.View(func(tx *bolt.Tx) error {
-		b := tx.Bucket([]byte("APIkeys"))
-		v := b.Get([]byte(key))
-		if len(v) == 0 {
-			return errors.New("key not found")
-		}
-		return nil
-	})
-	if err == nil {
-		return true
-	}
-	return false
-}
-
-func httpErrorf(w http.ResponseWriter, format string, a ...interface{}) {
-	err := fmt.Errorf(format, a...)
-	log.Println(err)
-	http.Error(w, err.Error(), http.StatusInternalServerError)
 }
